removed 128bit floats and hand-written logger (replaced with glog)

+++ tools/fc.cpp
@@ -7,14 +7,13 @@
 #include <stdexcept>
 #include <string>
 
-#include "quadmath.hpp"
 #include "file_io.hpp"
 #include "compute.hpp"
 #include "clean.hpp"
 
 int main(int argc, char ** argv) {
   try {
-    using float_t = __float128;
+    using float_t = double;
     if (argc != 2)
       throw std::invalid_argument("usage: fc <point_cloud_file>");
     auto * filename = argv[1];
